While documenting I realised the PlacementModel is using too much memory and is becoming the core structure that stores many millions of records so make its properties as small as possible from Int to UInt8.  So also changed where it was used thus many changes.

+++ Sources/CrozzleRectangleMaker/Calculators/EdgeCalculator.swift
@@ -49,9 +49,9 @@ public class EdgeCalculator {
                                         verticalPosition: UInt8(verticalLetter),
                                         verticalLength: UInt8(verticalWord.count),
                                         
-                                        width: width,
-                                        height: height,
-                                        score: score
+                                        width: UInt8(width),
+                                        height: UInt8(height),
+                                        score: UInt16(score)
                                     ))
                                 }
                             }
